requirements.txt
beautifulsoup4==4.13.4
discord.py==2.4.0
python-dotenv==1.1.0
PyYAML==6.0.2
Requests==2.32.3
